Add files via upload
more work needs to be done for view.java. that aside we need to move files into packages so that its more organized. So if you do run into problems you may need to comment out/remove the "package <name>" that's near the top b/c of the lack of packaging.