updated Song and created tracking
added a replayCount field to song to track play count, created tracking class. Have not tested if it works i'm stuck on how i should update the frequently played songs considering we need to go through albums and playlist; for each song count, grabbing the max, checking if it exists, and updating it.